Hw 30 (#172)
* init telegram module

* add telegram webhook

* add send confirm code for auth telegram bot

* add telegram bot auth confirmation

* add blog subscription flow

* add blog unsubscription flow

* add telegram bot notification flow

* post view query with sub blog

* refactor view blog query for public api to blog with subscription info

* refactor sub/unsub to blog flow add check sub status by user

* resolve conflict in blog repo

* refactor telegram flow

* remove import from blog sub use-case

+++ src/modules/users/entities/user.entity.ts
@@ -29,10 +29,10 @@ export class User {
   @Column({ type: 'enum', enum: CreateUserFormat })
   format: CreateUserFormat;
 
-  @Column({ type: 'int', nullable: true, default: null })
+  @Column({ type: 'bigint', nullable: true, default: null })
   telegram_id: number;
 
-  @Column({ type: 'int', nullable: true, default: null })
+  @Column({ type: 'bigint', nullable: true, default: null })
   chat_id: number;
 
   @OneToMany(() => Device, devices => devices.user)
